backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
pyyaml>=6.0
aiosqlite>=0.20.0
apscheduler>=3.10.0
cryptography>=42.0.0
bcrypt>=4.1.0
itsdangerous>=2.2.0
